Update patch set 1
Patch Set 1:

Hi Alex:
Can you ensure the src code won't be deleted, especially those in the private repos?
If not, it's better we have a local copy of the code, and with the extra information added, such as the commit id as you provided.
For example: the Contiv/VPP:
https://github.com/contiv/vpp/tree/master/vendor
or ovn-kubernetes:
https://github.com/openvswitch/ovn-kubernetes/tree/master/go-controller/vendor
These vendor dirs contains all the src code, not only the links

Patch-set: 1