Update patch set 1
Patch Set 1:

1. Yes, I would add some condition check before apply the change.
2.Yes, I had tested the clusterDNS setting of this patch, it really would affect the work of pod.
On the other side, there are some cases that the CoreDNS really can not work due to the /etc/resolv.conf setting(which is supported with local dnsmasq or systemd-resolved by NetworkManager) of the target system.
So we would investigate it further to get a better solution.

Patch-set: 1
Label: Verified=0