Update patch set 12
Patch Set 12:

> Build Failed
 > 
 > https://jenkins.akraino.org/job/iec-tox-verify-master/20/ : FAILURE
 > 
 > Logs: https://logs.akraino.org/releng/vex-yul-akraino-jenkins-prod-1/iec-tox-verify-master/20

Somehow the rebase didn't actually happen for this version of the patch. The error given by Jenkins is "ERROR: unknown environment 'coala'" which means it didn't take into consideration the last 2 merges in the repo. 

Now it seems everything is back on track, this patch can be merged

Patch-set: 12