Add PONSim test scripts for running SIAB
The test script is designed to be run from a CI job but it can be run
manually by providing the K8S master to connect to and the .kube config
file needed to access it.

The SIAB tests, which are part of opencord/cord-tester, are run inside a
Docker container, by default iecedge/cord-tester, multi-arch enabled.

Also add the Dockerfile used to build the cord-tester docker image
mentioned above.

Signed-off-by: Ciprian Barbu <ciprian.barbu@enea.com>
Change-Id: I17475ce90c152f7541ae806b99d6709c1675cda1