Update patch set 6
Patch Set 6:

I can write a short README, but the idea is pretty simple. You first run install.sh, on the K8S master and then you run test.sh from the jump host.
The purpose of these scripts is first and foremost to be run in CI. Have a look at the related change #1243, for ci-management

Patch-set: 6
Label: Verified=0