Update patch set 11
Patch Set 11:

This is the final form of this change for now. I had to push the changes to my personal repo in order to test them in Jenkins, in sandbox. The logs can be found here for now:
https://jenkins.akraino.org/sandbox/job/iec-test-seba_on_arm-fuel-virtual-ubuntu1604-daily-master/24/console

The actual robot tests fail because the ponsim-pod tosca loader fails to create the ATTServiceWorkflowDriverInstance, which is something we've seen happening on virtual pods, on baremetal the tests advance more, but still fail with other errors that only happen on arm64. We will need to fix those after this.

Patch-set: 11
Reviewer: Gerrit User 147 <147@4cada01b-0314-4fe0-84a1-f9ddc89863f9>
Label: Verified=0