Initial Commit to Add RestAPI Server to IEC
Initial commit to add a Rest API service(http server) to IEC,
which now just support simple Restful request as:
http://IP:port/v1/iec/status
and return a simple JSON reply.
It would be enhanced with more features in the future.

This server listens at TCP port 9131. It can be deployed as a Kubernetes
service, and accessed by http://svcIP/v1/iec/xxx, or by
http://nodeIP:31131/v1/iec/xxx.

Now the commit includes:
1. The API server implementation itself
2. The docker images related files
3. The Kubernetes deployment scripts
4. README.md
5. The Makefile to manage most things.

Signed-off-by: trevor tao <trevor.tao@arm.com>
Change-Id: I0891855a5182fc193fd26e7232903aac7e9b4d12
Signed-off-by: trevor tao <trevor.tao@arm.com>