Fix the bug of installation failure
When using the K8s with 1.13.0 version, it will result
deployment failed. It mainly because that parsing the
"kubeadm join" command in older version is different
from 1.15.2 version. When deploying the K8s with 1.13.0
by scripts, the IP address was consider as token key.
Finally, it results in the worker node installation failed.

Signed-off-by: Jingzhao.Ni <Jingzhao.Ni@arm.com>
Change-Id: I088392e416bf4ef92be78196a83987279d29e268