Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+1
> 
> Could you share what the issue is that caused this patch?

Sure. It's actually something I discovered last year, while working on the SiaB test framework. There are plenty of details in https://jira.akraino.org/browse/IEC-26, but briefly it's about an error with pwhash tests which appears while building the libsodium pip wheel. One of the negative tests seems to pass an invalid memory size which causes pwhash test to try and mmap an enormous amount of memory which in turn causes memory depletion and the test fails after a long time. With the default vm.overcommit_memory set to 1, the kernel does no additional checks during the mmap system call, while if we set it to 0 or 2 there are extra checks performed and the test passes.
I think this value for overcommit_memory is set as default for the Ubuntu Server aarch64 image, but I'm not entirely sure.

Patch-set: 2